in1n redefined as integer*1

+++ modules.f90
@@ -41,7 +41,7 @@ INTEGER maxcuantas
 
 integer curvature, radio
 
-integer*4, allocatable :: in1n(:,:,:,:)
+integer*1, allocatable :: in1n(:,:,:,:)
 integer*4, allocatable :: in1tmp(:)
 integer, allocatable ::  maxpos(:,:,:)
 integer, allocatable ::  minpos(:,:,:)
